+++ clients/desktop/src/vault/keysign/shared/KeysignTxOverview.tsx
@@ -1,5 +1,6 @@
 import { fromChainAmount } from '@core/chain/amount/fromChainAmount'
 import { Chain } from '@core/chain/Chain'
+import { formatFee } from '@core/chain/tx/fee/format/formatFee'
 import { getBlockExplorerUrl } from '@core/chain/utils/getBlockExplorerUrl'
 import { fromCommCoin } from '@core/mpc/types/utils/commCoin'
 import { KeysignPayload } from '@core/mpc/types/vultisig/keysign/v1/keysign_message_pb'
@@ -23,7 +24,6 @@ import { useMemo } from 'react'
 import { useTranslation } from 'react-i18next'
 
 import { nativeSwapChains } from '../../../chain/swap/native/NativeSwapChain'
-import { formatFee } from '../../../chain/tx/fee/utils/formatFee'
 import { useCopyTxHash } from '../../../chain/ui/hooks/useCopyTxHash'
 import { useCoinPriceQuery } from '../../../coin/query/useCoinPriceQuery'
 import { useFiatCurrency } from '../../../preferences/state/fiatCurrency'
