+++ clients/desktop/src/vault/swap/form/info/SwapFees.tsx
@@ -1,3 +1,4 @@
+import { formatFee } from '@core/chain/tx/fee/format/formatFee'
 import { VStack } from '@lib/ui/layout/Stack'
 import { MatchQuery } from '@lib/ui/query/components/MatchQuery'
 import { Text, TextColor } from '@lib/ui/text'
@@ -7,7 +8,6 @@ import { ComponentType, FC, PropsWithChildren } from 'react'
 import { useTranslation } from 'react-i18next'
 import styled from 'styled-components'
 
-import { formatFee } from '../../../../chain/tx/fee/utils/formatFee'
 import { Skeleton } from '../../../../components/skeleton'
 import { useSwapChainSpecificQuery } from '../../queries/useSwapChainSpecificQuery'
 import { useSwapFeesQuery } from '../../queries/useSwapFeesQuery'
