+++ core/chain/tx/fee/format/feeUnit.ts
@@ -3,7 +3,7 @@ import { chainFeeCoin } from '@core/chain/coin/chainFeeCoin'
 import { gwei } from '@core/chain/tx/fee/evm/gwei'
 import { isOneOf } from '@lib/utils/array/isOneOf'
 
-import { getUtxoFeeUnit } from '../../../utxo/fee/getUtxoFeeUnit'
+import { getUtxoFeeUnit } from '../utxo/getUtxoFeeUnit'
 
 export const getFeeUnit = (chain: Chain): string => {
   if (isOneOf(chain, Object.values(EvmChain))) {
